Refactor mock client response as subclass
Consistency with other implementations, easier extensibility
